+++ src/ImageHandler.ts
@@ -2,6 +2,9 @@ import {getSetting} from "./utils/settings";
 
 import ImageRequest from "./ImageRequest";
 import * as imageOps from "./image-ops";
+import RequestNotProcessedException from "./errors/RequestNotProcessedException";
+import {ImageExtensions, ParsedSchemaItem} from "./types/common";
+import {Sharp} from "sharp";
 
 export default class ImageHandler {
   private readonly request: ImageRequest;
@@ -20,44 +23,50 @@ export default class ImageHandler {
     // Get the original image
     const originalImageObject = this.request.originalImageObject
     const originalImageBody = this.request.originalImageBody
+    if (!originalImageBody || !originalImageObject || !this.request.sharpObject) {
+      throw new RequestNotProcessedException('Original image body or image object not available prior to processing.')
+    }
 
     let contentType = originalImageObject.ContentType
-
     let format
     let bufferImage
+
     // We have some edits to process
-    if (Object.keys(this.request.edits).length) {
+    if (this.request.edits && Object.keys(this.request.edits).length) {
       try {
         // We're calling rotate on this immediately in order to ensure metadata for rotation doesn't get lost
         const pipeline = this.request.sharpObject.rotate()
         await this.applyEdits(pipeline, this.request.edits)
         await this.applyOptimizations(pipeline)
         bufferImage = await pipeline.toBuffer()
+        // pipeline.options is not in the Sharp ts definitions. Should probably create a PR
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+        // @ts-ignore
         format = pipeline.options.formatOut
       } catch (err) {
         console.error('Unhandlable image encountered', err)
-        bufferImage = Buffer.from(originalImageBody, 'binary')
+        bufferImage = Buffer.from(originalImageBody.toString(), 'binary')
       }
     } else {
       // No edits, just return the original
-      bufferImage = Buffer.from(originalImageBody, 'binary')
+      bufferImage = Buffer.from(originalImageBody.toString(), 'binary')
     }
     if (format) {
       switch (format.toLowerCase()) {
-      case 'jpeg':
-      case 'jpg':
+      case ImageExtensions.JPEG:
+      case ImageExtensions.JPG:
         contentType = 'image/jpeg'
         break
-      case 'png':
+      case ImageExtensions.PNG:
         contentType = 'image/png'
         break
-      case 'webp':
+      case ImageExtensions.WEBP:
         contentType = 'image/webp'
         break
-      case 'gif':
+      case ImageExtensions.GIF:
         contentType = 'image/gif'
         break
-      case 'heif':
+      case ImageExtensions.HEIF:
         contentType = 'image/avif'
         break
       case 'input':
@@ -81,7 +90,7 @@ export default class ImageHandler {
    * @param {sharp} image - The original image.
    * @param {Object} edits - The edits to be made to the original image.
    */
-  async applyEdits(image, edits) {
+  async applyEdits(image: Sharp, edits: { [operation: string]: ParsedSchemaItem }) {
     await imageOps.restrictSize(image, this.request.originalMetadata)
     await imageOps.apply(image, edits)
   }
@@ -93,11 +102,16 @@ export default class ImageHandler {
    * @param headers
    * @returns {Promise<sharp>}
    */
-  async applyOptimizations(image) {
+  async applyOptimizations(image: Sharp) {
     // const minColors = 128 // arbitrary number
     // const maxColors = 256 * 256 * 256 // max colors in RGB color space
-    const {edits, headers} = this.request
-    const {auto} = edits
+    const {edits}: ImageRequest = this.request
+    if (!edits) {
+      // TODO: Is this safe?
+      return
+    }
+
+    const auto = edits.auto
 
     let autoVals = auto.processedValue
     if (!Array.isArray(autoVals)) {
@@ -106,16 +120,18 @@ export default class ImageHandler {
 
     // Determine our quality - if it was implicitly determined, we'll use the environment setting rather than the schema
     let quality = getSetting('DEFAULT_QUALITY')
-    if (edits.q.implicit !== true) {
-      quality = parseInt(edits.q.processedValue)
-      if (quality < 1) {
-        quality = 1
-      } else if (quality > 100) {
-        quality = 100
+    if (edits) {
+      if (!edits.q.implicit) {
+        quality = parseInt(edits.q.processedValue)
+        if (quality < 1) {
+          quality = 1
+        } else if (quality > 100) {
+          quality = 100
+        }
       }
     }
 
-    let fm = edits.fm.processedValue
+    const fm = edits.fm.processedValue
 
     if (autoVals.includes('compress')) {
       quality = getSetting('DEFAULT_COMPRESS_QUALITY')
@@ -123,7 +139,7 @@ export default class ImageHandler {
 
 
     // adjust quality based on file type
-    if (fm === 'jpg' || fm === 'jpeg') {
+    if (fm === ImageExtensions.JPG || fm === ImageExtensions.JPEG) {
       if (autoVals.includes('compress') && quality < 100 && edits.q !== undefined) {
         image.jpeg({
           quality: quality,
@@ -141,7 +157,8 @@ export default class ImageHandler {
       })
     } else if (fm === 'webp') {
       const options = {
-        quality: quality
+        quality: quality,
+        lossless: false
       }
       if ('lossless' in edits && edits.lossless.processedValue === true) {
         options.lossless = true
@@ -149,7 +166,8 @@ export default class ImageHandler {
       image.webp(options)
     } else if (fm === 'avif') {
       const options = {
-        quality: quality
+        quality: quality,
+        lossless: false
       }
       if ('lossless' in edits && edits.lossless.processedValue === true) {
         options.lossless = true